remove: drop `boot_diag.rs` example and update related configurations
- Removed the `boot_diag.rs` Linux-specific example to simplify codebase and maintenance.
- Adjusted CI documentation build command to use the `--workspace` flag for consistency.
- Minor cleanup in `VzBackend` to remove unnecessary cast for file descriptor.

+++ src/backend/vz/backend.rs
@@ -123,7 +123,7 @@ impl VzBackend {
                         return;
                     }
                     let fd = unsafe { (*connection).fileDescriptor() };
-                    let _ = tx_clone.send(Ok(fd as i32));
+                    let _ = tx_clone.send(Ok(fd));
                 },
             );
